+++ reactivex/operators/_timeout.py
@@ -9,11 +9,9 @@ from reactivex.disposable import (
     SingleAssignmentDisposable,
 )
 from reactivex.internal import curry_flip
-from reactivex.scheduler import EventLoopScheduler
+from reactivex.scheduler import TimeoutScheduler
 
 _T = TypeVar("_T")
-
-_default_timeout_scheduler = EventLoopScheduler()
 
 
 @curry_flip
@@ -50,7 +48,7 @@ def timeout_(
         observer: abc.ObserverBase[_T],
         scheduler_: abc.SchedulerBase | None = None,
     ) -> abc.DisposableBase:
-        _scheduler = scheduler or scheduler_ or _default_timeout_scheduler
+        _scheduler = scheduler or scheduler_ or TimeoutScheduler.singleton()
 
         switched = [False]
         _id = [0]